cleaned rest of utils.py; renamed to _utils.py; added module element_utils
