+++ src/spatialdata/element_utils/__init__.py
@@ -1,0 +1,11 @@
+from spatialdata.element_utils._utils import (
+    iterate_pyramid_levels,
+    multiscale_spatial_image_from_data_tree,
+    unpad_raster,
+)
+
+__all__ = [
+    "iterate_pyramid_levels",
+    "multiscale_spatial_image_from_data_tree",
+    "unpad_raster",
+]
