+++ src/spatialdata/element_utils/_utils.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-# from https://stackoverflow.com/a/24860799/3343783
-import re
 from collections.abc import Generator
-from typing import TYPE_CHECKING, Union
+from typing import Union
 
-import dask.array as da
 import numpy as np
+from dask import array as da
 from datatree import DataTree
 from multiscale_spatial_image import MultiscaleSpatialImage
 from spatial_image import SpatialImage
@@ -17,13 +15,6 @@ from spatialdata._core.spatialdata_operations import (
     set_transformation,
 )
 from spatialdata._core.transformations import Sequence, Translation
-from spatialdata._types import ArrayLike
-
-# I was using "from numbers import Number" but this led to mypy errors, so I switched to the following:
-Number = Union[int, float]
-
-if TYPE_CHECKING:
-    pass
 
 
 def unpad_raster(raster: Union[SpatialImage, MultiscaleSpatialImage]) -> Union[SpatialImage, MultiscaleSpatialImage]:
@@ -70,6 +61,8 @@ def unpad_raster(raster: Union[SpatialImage, MultiscaleSpatialImage]) -> Union[S
     translation_values: list[float] = []
     unpadded = raster
 
+    # TODO: this "if else" will be unnecessary once we remove the concept of intrinsic coordinate systems and we make the
+    #  transformations and xarray coordinates more interoperable
     if isinstance(unpadded, SpatialImage):
         for ax in axes:
             if ax != "c":
@@ -144,35 +137,3 @@ def iterate_pyramid_levels(image: MultiscaleSpatialImage) -> Generator[DataArray
         assert len(v) == 1
         xdata = next(iter(v))
         yield xdata
-
-
-def _parse_list_into_array(array: Union[list[Number], ArrayLike]) -> ArrayLike:
-    if isinstance(array, list):
-        array = np.array(array)
-    if array.dtype != float:
-        array = array.astype(float)
-    return array
-
-
-def atoi(text: str) -> Union[int, str]:
-    return int(text) if text.isdigit() else text
-
-
-# from https://stackoverflow.com/a/5967539/3343783
-def natural_keys(text: str) -> list[Union[int, str]]:
-    """
-    alist.sort(key=natural_keys) sorts in human order
-    http://nedbatchelder.com/blog/200712/human_sorting.html
-    (See Toothy's implementation in the comments)
-    """
-    return [atoi(c) for c in re.split(r"(\d+)", text)]
-
-
-def _affine_matrix_multiplication(matrix: ArrayLike, data: ArrayLike) -> ArrayLike:
-    assert len(data.shape) == 2
-    assert matrix.shape[1] - 1 == data.shape[1]
-    vector_part = matrix[:-1, :-1]
-    offset_part = matrix[:-1, -1]
-    result = data @ vector_part.T + offset_part
-    assert result.shape[0] == data.shape[0]
-    return result  # type: ignore[no-any-return]
